Chrome Testcases Addes

+++ se/src/test/java/okw/SeInputField/SeInputField_Chrome_Test.java
@@ -35,9 +35,10 @@ public class SeInputField_Chrome_Test extends SeInputField_Test
     public void FirefoxAfter() throws Exception
     {
     	Runtime rt = Runtime.getRuntime();
-    	  if (System.getProperty("os.name").toLowerCase().indexOf("windows") > -1) 
-    	     rt.exec("taskkill Chrome");
-    	   else
-    	     rt.exec("pkill -f Chrome");
+    	
+    	if (System.getProperty("os.name").toLowerCase().indexOf("windows") > -1) rt.exec("taskkill Chrome");
+    	else rt.exec("pkill -f Chrome");
+    	  
+    	Thread.sleep( 1000 );
     }
 }
