+++ se/src/test/java/okw/SeRadioList/SeRadioList_Chrome_Test.java
@@ -28,9 +28,10 @@ public class SeRadioList_Chrome_Test extends SeRadioList_Test
     public void FirefoxAfter() throws Exception
     {
     	Runtime rt = Runtime.getRuntime();
-    	  if (System.getProperty("os.name").toLowerCase().indexOf("windows") > -1) 
-    	     rt.exec("taskkill Chrome");
-    	   else
-    	     rt.exec("pkill -f Chrome");
+    	
+    	if (System.getProperty("os.name").toLowerCase().indexOf("windows") > -1) rt.exec("taskkill Chrome");
+    	else rt.exec("pkill -f Chrome");
+    	  
+    	Thread.sleep( 1000 );
     }
 }
